Fix bug and refactor code to DeadlineStorage
