+++ src/main/java/seedu/duke/Duke.java
@@ -13,6 +13,7 @@ public class Duke {
      */
     private static DataReader dataReader = new DataReader();
     private static Storage storage = new Storage();
+    private static DeadlineStorage deadlineStorage = new DeadlineStorage();
 
     private static BudgetPlanner budgetPlanner = new BudgetPlanner();
     private static UI ui = new UI();
@@ -26,13 +27,13 @@ public class Duke {
         ArrayList<University> universities = dataReader.getUniversities();
         ArrayList<Module> allModules = dataReader.getModules();
         ArrayList<Module> modules = storage.getModules();
-        ArrayList<Deadline> deadlines = storage.getDeadlines();
+        ArrayList<Deadline> deadlines = deadlineStorage.getDeadlines();
         ui.printGreetingMessage();
-        storage.compareDeadlines(deadlines);
+        deadlineStorage.compareDeadlines(deadlines);
         while (isContinue) {
             userInput = in.nextLine();
             Command command = parser.parseUserCommand(userInput, universities, modules, allModules, storage,
-                    budgetPlanner, deadlines);
+                    deadlineStorage, budgetPlanner, deadlines);
             command.execute();
             isContinue = !command.getIsExit();
         }
