+++ src/main/java/seedu/duke/Storage.java
@@ -5,30 +5,19 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
-import java.util.concurrent.TimeUnit;
 
 public class Storage implements DatabaseInterface {
 
     private static final String SAVED_MODULES_FILE_PATH = "data/saved_modules.txt";
-    private static final String SAVED_DEADLINES_FILE_PATH = "data/deadlines.txt";
     private ArrayList<Module> modules;
-    private ArrayList<Deadline> deadlines;
 
     public Storage() {
         this.modules = new ArrayList<>();
-        this.deadlines = new ArrayList<>();
         try {
             initialiseDatabase();
         } catch (IOException e) {
             System.out.println("Initialise Saved Modules Failure");
-        }
-        try {
-            initialiseDeadlinesDatabase();
-        } catch (IOException e) {
-            System.out.println("Initialise Deadlines Failure");
         }
     }
 
@@ -80,18 +69,6 @@ public class Storage implements DatabaseInterface {
         return false;
     }
 
-    public void initialiseDeadlinesDatabase() throws IOException {
-        File savedDeadlinesFile = new File(SAVED_DEADLINES_FILE_PATH);
-        if (!savedDeadlinesFile.exists()) {
-            File directory = new File("data");
-            directory.mkdirs();
-            savedDeadlinesFile.createNewFile();
-            return;
-        }
-        readDeadlineData(SAVED_DEADLINES_FILE_PATH, deadlines);
-    }
-
-
     private void readModData(String modulesFilePath, ArrayList<Module> modules) {
         try (BufferedReader br = new BufferedReader(new FileReader(modulesFilePath))) {
             String line;
@@ -115,20 +92,6 @@ public class Storage implements DatabaseInterface {
                 } catch (NumberFormatException e) {
                     UI.printModuleCorruptDeleteMessage();
                 }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-
-    private void readDeadlineData(String modulesDeadlinePath, ArrayList<Deadline> deadlines) {
-        try (BufferedReader br = new BufferedReader(new FileReader(modulesDeadlinePath))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                String[] row = line.split("//");
-                Deadline deadline = new Deadline(row[0], row[1]);
-                deadlines.add(deadline);
             }
         } catch (IOException e) {
             e.printStackTrace();
@@ -211,32 +174,6 @@ public class Storage implements DatabaseInterface {
         return true;
     }
 
-    /**
-     * Deletes the deadline specified by user. Deadline will the removed from user's
-     * list of deadlines.
-     *
-     * @param indexToDelete Index of that deadline that is given in user input.
-     * @param deadlines     The ArrayList of deadlines.
-     * @param database      Database of the user's list of deadlines.
-     * @return True if successfully deleted the module, false if unsuccessful.
-     */
-    public static boolean deleteDeadline(int indexToDelete, ArrayList<Deadline> deadlines,
-                                         Storage database) {
-        int indexToZeroBased = indexToDelete - 1;
-        try {
-            deadlines.remove(indexToZeroBased);
-        } catch (IndexOutOfBoundsException e) {
-            UI.printDeleteNumError();
-            return false;
-        }
-        try {
-            database.writeDeadlinesToFile(deadlines);
-        } catch (IOException e) {
-            UI.printWriteToDatabaseFailureMessage();
-            return false;
-        }
-        return true;
-    }
 
     /**
      * Adds and overwrites ArrayList of user's saved modules list in database.
@@ -255,79 +192,11 @@ public class Storage implements DatabaseInterface {
     }
 
     /**
-     * Adds and overwrites ArrayList of user's deadlines in database.
-     *
-     * @param deadlines ArrayList of deadlines to be written into database.
-     * @throws IOException If input/output operations fail or are interrupted.
-     */
-    public void writeDeadlinesToFile(ArrayList<Deadline> deadlines) throws IOException {
-        FileWriter fw = new FileWriter(SAVED_DEADLINES_FILE_PATH);
-        String stringToAdd = "";
-        for (Deadline deadline : deadlines) {
-            stringToAdd += writeTaskPreparation(deadline.toString());
-        }
-        fw.write(stringToAdd);
-        fw.close();
-    }
-
-    /**
      * Returns list of modules in ArrayList type.
      *
      * @return ArrayList of modules.
      */
     public ArrayList<Module> getModules() {
         return modules;
-    }
-
-    public ArrayList<Deadline> getDeadlines() {
-        return deadlines;
-    }
-
-    public void addDeadlineToDeadlines(Deadline deadlineToAdd) {
-        if (deadlineToAdd == null) {
-            UI.printAddDeadlineFailureMessage();
-            return;
-        }
-        deadlines.add(deadlineToAdd);
-        try {
-            saveDeadlineToStorage(deadlineToAdd.toString());
-        } catch (IOException e) {
-            UI.printAddDeadlineFailureMessage();
-        }
-    }
-
-    private void saveDeadlineToStorage(String saveDeadlineString) throws IOException {
-        FileWriter fw = new FileWriter(SAVED_DEADLINES_FILE_PATH, true);
-        fw.write(writeTaskPreparation(saveDeadlineString));
-        fw.close();
-    }
-
-    public void compareDeadlines(ArrayList<Deadline> deadlines) {
-        Date date = new Date();
-        SimpleDateFormat formatter = new SimpleDateFormat("dd-MM-yyyy");
-        String todayDate = formatter.format(date);
-        int counter = 1;
-        try {
-            boolean hasReminderMsgPrinted = false;
-            for (Deadline deadline : deadlines) {
-                Date today = formatter.parse(todayDate);
-                Date deadlineDue = formatter.parse(deadline.getDueDate());
-                long timeDiff = Math.abs(deadlineDue.getTime() - today.getTime());
-                long daysDiff = TimeUnit.DAYS.convert(timeDiff, TimeUnit.MILLISECONDS);
-                if (daysDiff <= 7) {
-                    if (!hasReminderMsgPrinted) {
-                        UI.printReminderMessage();
-                        hasReminderMsgPrinted = true;
-                    }
-                    UI.printReminderDeadline(deadline, counter);
-                    counter++;
-                }
-            }
-            if (hasReminderMsgPrinted) {
-                UI.printLine();
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
     }
 }
