+++ src/main/java/seedu/duke/command/DeleteDeadlineCommand.java
@@ -1,30 +1,30 @@
 package seedu.duke.command;
 
 import seedu.duke.Deadline;
-import seedu.duke.Storage;
+import seedu.duke.DeadlineStorage;
 
 import java.util.ArrayList;
 
 public class DeleteDeadlineCommand extends Command {
 
-    private Storage storage;
+    private DeadlineStorage deadlineStorage;
 
     private int indexToRemove;
 
     private ArrayList<Deadline> deadlines;
 
-    public DeleteDeadlineCommand(Storage storage, int indexToRemove, ArrayList<Deadline> deadlines) {
-        this.storage = storage;
+    public DeleteDeadlineCommand(DeadlineStorage deadlineStorage, int indexToRemove, ArrayList<Deadline> deadlines) {
+        this.deadlineStorage = deadlineStorage;
         this.indexToRemove = indexToRemove;
         this.deadlines = deadlines;
     }
 
     @Override
     public void execute() {
-        boolean isDeleteSuccessful = Storage.deleteDeadline(indexToRemove, deadlines, storage);
+        boolean isDeleteSuccessful = DeadlineStorage.deleteDeadline(indexToRemove, deadlines, deadlineStorage);
         if (!isDeleteSuccessful) {
             return;
         }
-        ui.printDeleteModMessage();
+        ui.printDeleteDeadlineMessage();
     }
 }
