+++ src/test/java/seedu/duke/command/DeleteModuleCommandTest.java
@@ -3,6 +3,7 @@ package seedu.duke.command;
 import org.junit.jupiter.api.Test;
 import seedu.duke.DataReader;
 import seedu.duke.Deadline;
+import seedu.duke.DeadlineStorage;
 import seedu.duke.Module;
 import seedu.duke.Parser;
 import seedu.duke.Storage;
@@ -16,7 +17,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class DeleteModuleCommandTest {
     Storage storage = new Storage();
+    DeadlineStorage deadlineStorage = new DeadlineStorage();
     DataReader dataReader = new DataReader();
+
     @Test
     void userInput_remove_correctVariableTypeSuccess() {
         String userInput = "remove KU/1";
@@ -29,8 +32,8 @@ public class DeleteModuleCommandTest {
         modules.add(module);
         Parser parser = new Parser();
         BudgetPlanner budgetPlanner = new BudgetPlanner();
-        assertTrue(parser.parseUserCommand(userInput, universities, modules, puModules, storage, budgetPlanner,
-                deadlines) instanceof DeleteModuleCommand);
+        assertTrue(parser.parseUserCommand(userInput, universities, modules, puModules, storage, deadlineStorage,
+                budgetPlanner, deadlines) instanceof DeleteModuleCommand);
     }
 
     @Test
@@ -39,7 +42,7 @@ public class DeleteModuleCommandTest {
         Module module = new Module(1, "AE320", "Aerodynamics II", 3,
                 "ME4231", "Aerodynamics", 4);
         uniList1.add(module);
-        Storage.deleteModule(1, uniList1, storage,1);
+        Storage.deleteModule(1, uniList1, storage, 1);
         assertEquals(0, uniList1.size());
     }
 }
