+++ src/test/java/seedu/duke/command/ExitCommandTest.java
@@ -2,6 +2,7 @@ package seedu.duke.command;
 
 import org.junit.jupiter.api.Test;
 import seedu.duke.Deadline;
+import seedu.duke.DeadlineStorage;
 import seedu.duke.Module;
 import seedu.duke.Parser;
 import seedu.duke.Storage;
@@ -25,10 +26,11 @@ public class ExitCommandTest {
         ArrayList<Module> puModules = new ArrayList<>();
         ArrayList<Deadline> deadlines = new ArrayList<>();
         Storage storage = new Storage();
+        DeadlineStorage deadlineStorage = new DeadlineStorage();
         Parser parser = new Parser();
         BudgetPlanner budgetPlanner = new BudgetPlanner();
-        assertTrue(parser.parseUserCommand(userInput, universities, modules, puModules, storage, budgetPlanner,
-                deadlines) instanceof ExitCommand);
+        assertTrue(parser.parseUserCommand(userInput, universities, modules, puModules, storage, deadlineStorage,
+                budgetPlanner, deadlines) instanceof ExitCommand);
     }
 
     @Test
