+++ src/test/java/seedu/duke/command/HelpCommandTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import seedu.duke.Deadline;
+import seedu.duke.DeadlineStorage;
 import seedu.duke.Module;
 import seedu.duke.Parser;
 import seedu.duke.Storage;
@@ -37,10 +38,11 @@ public class HelpCommandTest {
         ArrayList<Module> puModules = new ArrayList<>();
         ArrayList<Deadline> deadlines = new ArrayList<>();
         Storage storage = new Storage();
+        DeadlineStorage deadlineStorage = new DeadlineStorage();
         Parser parser = new Parser();
         BudgetPlanner budgetPlanner = new BudgetPlanner();
-        assertTrue(parser.parseUserCommand(userInput, universities, modules, puModules, storage, budgetPlanner,
-                deadlines) instanceof HelpCommand);
+        assertTrue(parser.parseUserCommand(userInput, universities, modules, puModules, storage, deadlineStorage,
+                budgetPlanner, deadlines) instanceof HelpCommand);
     }
 
     //@@author dfa-reused
